// quadriga-fe/src/app/models/car.ts
export class Car {

    id: number = 0;
    available: boolean = true;
    year: number = 0;
    mileageCounter: number = 0;
    seatingCapacity: number = 0;
    plate: string = '';
    description: string = '';
    price: number = 0;
    colorId: number = 0;
    categoryId: number = 0;
    modelId: number = 0;
 //   brandId: number = 0;
    fuelTypeId: number = 0;
    transmissionTypeId: number = 0;
    createdDate: Date = new Date();
    updatedDate: Date = new Date();
}